requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
python-telegram-bot==20.2
requests==2.28.2
SQLAlchemy==2.0.9
